Utils -bits count

+++ utils/utils.cpp
@@ -1,13 +1,13 @@
 #include <cmath>
 #include <iostream>
 
-int calcNumBits(int upperBound) {
+int calcNumBits(double upperBound) {
 
     if (upperBound <= 0) {
         return 1;
     }
 
-    int numBits = std::ceil(std::log2(static_cast<double>(upperBound) + 1));
+    int numBits = std::ceil(std::log2(upperBound + 1));
 
     return numBits;
 }
